Rename directories
- Rename pkg/api/projectedresource to pkg/api/sharedresource
- Rename arguments for update/generated.sh to use new name
- Regenerate informers and listers
- Rename golang package imports
- Update controller with new informer factory name
- Update e2e tests to use new client
