+++ pkg/client/listers.go
@@ -1,7 +1,7 @@
 package client
 
 import (
-	sharev1alpha1 "github.com/openshift/csi-driver-projected-resource/pkg/generated/listers/projectedresource/v1alpha1"
+	sharev1alpha1 "github.com/openshift/csi-driver-projected-resource/pkg/generated/listers/sharedresource/v1alpha1"
 	corev1 "k8s.io/client-go/listers/core/v1"
 )
