+++ test/framework/share.go
@@ -7,7 +7,7 @@ import (
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	shareapi "github.com/openshift/csi-driver-projected-resource/pkg/api/projectedresource/v1alpha1"
+	shareapi "github.com/openshift/csi-driver-projected-resource/pkg/api/sharedresource/v1alpha1"
 )
 
 func CreateShare(t *TestArgs) {
@@ -25,7 +25,7 @@ func CreateShare(t *TestArgs) {
 			},
 		},
 	}
-	_, err := shareClient.ProjectedresourceV1alpha1().Shares().Create(context.TODO(), share, metav1.CreateOptions{})
+	_, err := shareClient.SharedresourceV1alpha1().Shares().Create(context.TODO(), share, metav1.CreateOptions{})
 	if err != nil && !kerrors.IsAlreadyExists(err) {
 		t.T.Fatalf("error creating test share: %s", err.Error())
 	}
@@ -45,7 +45,7 @@ func CreateShare(t *TestArgs) {
 				},
 			},
 		}
-		_, err := shareClient.ProjectedresourceV1alpha1().Shares().Create(context.TODO(), share, metav1.CreateOptions{})
+		_, err := shareClient.SharedresourceV1alpha1().Shares().Create(context.TODO(), share, metav1.CreateOptions{})
 		if err != nil && !kerrors.IsAlreadyExists(err) {
 			t.T.Fatalf("error creating test share: %s", err.Error())
 		}
@@ -59,13 +59,13 @@ func ChangeShare(t *TestArgs) {
 		name = t.SecondName
 	}
 	t.T.Logf("%s: start change share %s", time.Now().String(), name)
-	share, err := shareClient.ProjectedresourceV1alpha1().Shares().Get(context.TODO(), name, metav1.GetOptions{})
+	share, err := shareClient.SharedresourceV1alpha1().Shares().Get(context.TODO(), name, metav1.GetOptions{})
 	if err != nil {
 		t.T.Fatalf("error getting share %s: %s", name, err.Error())
 	}
 	share.Spec.BackingResource.Kind = "Secret"
 	share.Spec.BackingResource.Name = "pull-secret"
-	_, err = shareClient.ProjectedresourceV1alpha1().Shares().Update(context.TODO(), share, metav1.UpdateOptions{})
+	_, err = shareClient.SharedresourceV1alpha1().Shares().Update(context.TODO(), share, metav1.UpdateOptions{})
 	if err != nil {
 		t.T.Fatalf("error updating share %s: %s", name, err.Error())
 	}
@@ -78,7 +78,7 @@ func DeleteShare(t *TestArgs) {
 		name = t.ShareToDelete
 	}
 	t.T.Logf("%s: start delete share %s", time.Now().String(), name)
-	err := shareClient.ProjectedresourceV1alpha1().Shares().Delete(context.TODO(), name, metav1.DeleteOptions{})
+	err := shareClient.SharedresourceV1alpha1().Shares().Delete(context.TODO(), name, metav1.DeleteOptions{})
 	if err != nil && !kerrors.IsNotFound(err) {
 		t.T.Fatalf("error deleting share %s: %s", name, err.Error())
 	}
